moving tests to different file
